+++ VNC1L_BOMS.h
@@ -27,10 +27,10 @@ class VNC1L_BOMS
     NewSoftSerial _vnc1l;
     byte _pin_rx;
     byte _pin_tx;
-    int _baud;
+    long _baud;
     
   public:
-    VNC1L_BOMS(int baud, byte pin_rx, byte pin_tx);
+    VNC1L_BOMS(long baud, byte pin_rx, byte pin_tx);
     void file_open(const String &file);
     void file_seek(long offset);
     void file_read(int count, byte[]);
